增加api环境
